ShapeCheck was reporting a file clean without reading a line of it
`CSharpParseOptions` with no preprocessor symbols treats every `#if X` region
as DISABLED TEXT. It is not parsed, so it cannot produce a syntax error, and
the file still counts towards "checked N files". `Game/OnnxSpeech.cs` is 380
lines inside `#if LEDGER_ONNX`, and the footer has been saying
"0 shape errors (169 files)" while having read none of them.

That is rule 3b wearing a compiler's clothes: a zero that cannot tell "nothing
wrong" from "nothing examined", in the instrument whose whole job is to catch
what only Windows CI can otherwise see.

IT FOUND A REAL ERROR ON THE FIRST RUN. `CS1503: cannot convert from 'long' to
'uint'` — `int * uint` widens to long in C#, because uint cannot hold a
negative int, so the cast had to go before the multiply rather than around it.
Seconds here against a ~28-minute round trip, on a file nothing had compiled.

Symbols are `LEDGER_ONNX`, `UNITY_EDITOR`, `UNITY_STANDALONE_WIN`, and the
count of files carrying conditional code goes into the footer beside the error
count — so a symbol quietly dropping out shows up as a number moving rather
than as continued silence. Two files today.

WHAT THIS DOES NOT DO, said plainly. ShapeCheck is reference-independent by
design: it keeps only the diagnostics that need no assemblies. It has now READ
the backend and can speak for its syntax and for the semantic family in its
allow-list. It cannot resolve `Microsoft.ML.OnnxRuntime`, so anything about
those types is still a question only a build answers.
0 lint errors, 0 shape errors (169 files, 2 with conditional code), 0 shadowed Core types, 13 tool project(s) tracked, 35 on the reach ledger, shape ok (clips, barks, manifests), voice cast ok (4 uncast principal(s)), voice-gen ok (25 checks, 2010-line batch), voice-live ok (178 checks), slop 91/91 (19 patterns, 3933 strings), 13 card-writing rules, 60 cards shipped as edited, 40 probe calls staged, 17 queue items ready, docs 52/52 clean, 10 attribution check(s), Game layer compiles (166 files), 0 nested-type errors (240 Core types), 0 static/instance errors (72 members, 417 bodies), 0 filename-as-type errors (169 files, 13 filenames that are not types), 0 namespace-as-value errors (169 files, 4 segments in scope), workflow steps ok (1014 under the dispatch ceiling), 26 frame-drift checks (0 failed), 978 verdict keys, 18 new (run --learn), verdict format ok (selftest + newest run), 29 save-chaos checks, 3 soak checks (500 days x2), 84 adversary checks, 0 stale anchors, 3490 CoreTests.

+++ ledger/Assets/Scripts/Game/OnnxSpeech.cs
@@ -315,7 +315,12 @@ namespace Ledger.Game
                 int wav = (h - promptMels) * SamplesPerMel;
                 if (wav <= 0) { Why = "the prompt is longer than the line"; return null; }
 
-                var seed = new Gauss(unchecked(tokens.Length * 2654435761u
+                // `int * uint` widens to LONG in C#, because uint cannot hold
+                // a negative int — so this needs the cast before the multiply
+                // rather than around it. Caught by ShapeCheck in seconds once
+                // it was told to read conditional code, which is the whole
+                // argument for having done that.
+                var seed = new Gauss(unchecked((uint)tokens.Length * 2654435761u
                                                + (uint)promptTokens));
                 var z = new DenseTensor<float>(new[] { 1, 80, h });
                 for (int i = 0; i < 80 * h; i++) z.SetValue(i, seed.Next());
